Move build_feedback_for_auto_test + artifact-recovery match (parent #680)
Continue the meta-issue #680 reduction effort. Move two more clusters
out of `turn.rs` into existing sibling modules — no new module created.

Migrated to `feedback_builders.rs` (`pub(super)`):

* `build_feedback_for_auto_test(&AutoTestPlan, &AutoTestResult, &Path,
    changed_files) -> FeedbackFrame` — the auto-test FeedbackFrame
    builder. Already had multiple external callers in
    `verifier_skill.rs` (4 sites) and `verifier_orchestration.rs` (2
    sites), so it fits the existing feedback-builders cluster.

`feedback_builders.rs` imports gain
`super::auto_test::{AutoTestPlan, AutoTestResult, classify_auto_test}`.

Migrated to `task_contract.rs` (`pub(super)` + private):

* `pub(super) fn repo_edit_satisfies_artifact_recovery_target(category,
    relative_path, target) -> bool` — does a repo edit satisfy the
    active `RecoveryTarget`? Naturally belongs with `RecoveryTarget`,
    `ArtifactRole`, and `role_from_repo_edit` SSOTs.
* `fn test_artifact_path_family(path) -> Option<&'static str>` —
    module-private supporting helper, classifies a path into a known
    test-artifact family (rust-integration / pytest / typescript-test
    / javascript-test).

`turn.rs` adjustments:

* Removed the 3 fn definitions (originally at lines 553-654).
* Updated 1 production caller of `repo_edit_satisfies_artifact_recovery_target`
  to use the new `super::task_contract::*` path; the existing
  `impl Agent::repo_edit_satisfies_artifact_recovery_target` veneer
  shell stays as the entry point.
* Test mod imports rewrite:
  - `repo_edit_satisfies_artifact_recovery_target` moved from
    `super::{...}` to
    `super::super::task_contract::repo_edit_satisfies_artifact_recovery_target`
    (via the `task_contract` group import).
  - `build_feedback_for_auto_test` test caller rewrites
    `super::build_feedback_for_auto_test` →
    `super::super::feedback_builders::build_feedback_for_auto_test`.
* Cleaned up `super::auto_test::{...}` import block (removed
  `AutoTestPlan` / `AutoTestResult` / `classify_auto_test` — now only
  used in tests via `super::auto_test::*`).
* Gated `FeedbackFrameDraft` / `build_feedback_frame` imports with
  `#[cfg(test)]` since they're only used in test mod after the
  `build_feedback_for_auto_test` migration.

`verifier_skill.rs` / `verifier_orchestration.rs` adjustments:

* Sweep `super::turn::build_feedback_for_auto_test` →
  `super::feedback_builders::build_feedback_for_auto_test`
  (6 caller sites total).

LOC delta:

* `turn.rs`: 28,496 → 28,402 (-94 LOC)
* `feedback_builders.rs`: 201 → 240 (+39 LOC)
* `task_contract.rs`: + ~60 LOC
* Cumulative `turn.rs`: 39,489 → 28,402 (-11,087 LOC, -28.1%)

`pub(super)` matching pre-extraction visibility. No facade re-export
(DR3-001). `turn.rs` no longer hosts the moved helpers.

Verification:

* `cargo fmt`
* `cargo clippy --lib --all-targets -- -D warnings` clean
* `cargo test --lib` 3,114 passed / 0 failed
